Trying to create an AI based on concepts objects
The concept are composed by a name and associated words. For example,
the concept of "me" is composed of words "I, me, my, mine".
Once I extracted the concepts of a sentence, I can create connections
between them to formatate sentences.

It's a console based program written in C++ right now.